[PRED-16] 플레이어 유닛이 집단 이동하다가 웨이브 시작 시 개별 행동 하도록 수정, 스킬 사거리 적용.

+++ InWanderLand/Unit_TemplateData.h
@@ -29,7 +29,7 @@ namespace application
 
 			/// Decreasing Damage Elements
 			int m_defensePoint;
-			int m_dodgeProbability;					// 회피율
+			float m_dodgeProbability;					// 회피율
 			float m_criticalDamageDecreaseMultiplier;	// 피격 시 치명타 피해 감소 배율
 
 			float m_idRadius;
